pyproject.toml
[project]
name = "task-management-system"
version = "0.1.0"
description = ""
authors = [
    {name = "hdbv95", email = "hdbv95@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"

[tool.poetry]
packages = [
    { include = "task_manager" }
]

[tool.poetry.dependencies]
asgiref = "==3.8.1"
colorama = "==0.4.6"
coverage = "==7.6.12"
django = "==5.1.6"
django-filter = "==25.1"
djangorestframework = "==3.15.2"
djangorestframework-simplejwt = "==5.4.0"
docopt = "==0.6.2"
drf-yasg = "==1.21.8"
inflection = "==0.5.1"
iniconfig = "==2.0.0"
packaging = "==24.2"
pluggy = "==1.5.0"
psycopg2-binary = "==2.9.10"
pyjwt = "==2.10.1"
pytest = "==8.3.4"
pytest-cov = "==6.0.0"
pytest-django = "==4.10.0"
pytest-watch = "==4.2.0"
pytz = "==2025.1"
pyyaml = "==6.0.2"
sqlparse = "==0.5.3"
tzdata = "==2025.1"
uritemplate = "==4.1.1"
watchdog = "==6.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"
black = "^25.1.0"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
